requirements.txt
asgiref==3.3.1
coverage==5.3.1
Django==3.1.5
factory-boy==3.2.0
Faker==5.3.0
fontawesome-free==5.15.1
Pillow==8.1.0
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
